+++ src/Common/Result/Contacts/MobilePhoneItemResult.php
@@ -5,9 +5,11 @@ declare(strict_types=1);
 namespace B24io\Loyalty\SDK\Common\Result\Contacts;
 
 use B24io\Loyalty\SDK\Common\VerificationStatus;
+use B24io\Loyalty\SDK\Core\Exceptions\InvalidArgumentException;
 use B24io\Loyalty\SDK\Core\Result\AbstractItem;
 use Exception;
 use libphonenumber\PhoneNumber;
+use libphonenumber\PhoneNumberUtil;
 
 
 /**
@@ -24,7 +26,11 @@ class MobilePhoneItemResult extends AbstractItem
     {
         switch ($offset) {
             case 'number':
-                return $this->phoneNumberUtil->parse($this->data[$offset], null);
+                $phoneNumberUtil = PhoneNumberUtil::getInstance();
+                if ($phoneNumberUtil === null) {
+                    throw new InvalidArgumentException('cannot create libphonenumber util instance');
+                }
+                return $phoneNumberUtil->parse($this->data[$offset], null);
             case 'verificationStatus':
                 return new VerificationStatus($this->data['verification_status']);
             default: